update view list store, list product

+++ app/src/main/java/gem/training3/enterprisenetwork/screen/allproduct/AllProductPresenter.java
@@ -6,5 +6,5 @@ import gem.training3.enterprisenetwork.base.BasePresenter;
  * Created by huylv on 26-Feb-16.
  */
 public interface AllProductPresenter extends BasePresenter {
-    void getProductByStoreId(int storeId);
+    void getProductByStoreId(Long storeId);
 }
